Game launches now and displays two room objects. Added a lot more code on the game mechanics side of things.

+++ src/world/Room.cpp
@@ -3,7 +3,7 @@
 #include "src/entities/Entity.h"
 #include "src/doodad/Doodad.h"
 
-Room::Room(int width, int depth) {
+Room::Room(float width, float depth) {
     this->width = width;
     this->depth = depth;
 } // constructor
